User commit

+++ src/main/java/com/example/demo/repository/ResourceAllocationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.ResourceAllocation;
 
 @Repository
-public interface ResourceAllocationRepository extends JpaRepository<ResourceAllocation,Integer>{
+public interface ResourceAllocationRepository extends JpaRepository<ResourceAllocation,Long>{
        
 }
